Workstation should manually cache and invalidate audio in memory as needed by craft/dub ahead (#70)
- Drop Caffeine in favor of manual solution
- On Craft, submit audio caching requests in order of the appearance of the audio in the segment from first to last
- On Dub, always check to ensure that all audio required by the mixer frame are already cached, and if they are not, skip the dub cycle
- On Craft, Invalidate all audio that is not currently or future necessitated
- Telemetry keeps track of the amount of time spent on each task versus total

https://www.pivotaltracker.com/story/show/186556102

+++ gui/src/main/java/io/xj/gui/services/FabricationServiceImpl.java
@@ -161,7 +161,7 @@ public class FabricationServiceImpl implements FabricationService {
       .setInputTemplateKey(inputTemplateKey.get())
       .setOutputChannels(Integer.parseInt(outputChannels.get()))
       .setOutputFileMode(outputFileMode.get())
-      .setOutputFrameRate(Double.parseDouble(outputFrameRate.get()))
+      .setOutputFrameRate(Integer.parseInt(outputFrameRate.get()))
       .setOutputMode(outputMode.get())
       .setOutputPathPrefix(outputPathPrefix.get())
       .setOutputSeconds(Integer.parseInt(outputSeconds.get()));
